requirements.txt
pytest==5.4.3
eagerpy==0.27.0
numpy==1.19.0
torch==1.4.0
jaxlib==0.1.51
jax==0.1.72
tensorflow-cpu==2.2.0
